consolidated and updated the prompts
