+++ packages/agent-toolkit/src/core/tools/platform-api-tools/index.ts
@@ -69,8 +69,7 @@ import { GetAgentTool } from './agents-tools/get-agent/get-agent-tool';
 import { CreateAgentTool } from './agents-tools/create-agent/create-agent-tool';
 import { DeleteAgentTool } from './agents-tools/delete-agent/delete-agent-tool';
 import { ListWorkflowsTool } from './workflows-tools/list-workflows/list-workflows-tool';
-import { ActivateWorkflowTool } from './workflows-tools/activate-workflow/activate-workflow-tool';
-import { DeactivateWorkflowTool } from './workflows-tools/deactivate-workflow/deactivate-workflow-tool';
+import { SetWorkflowActiveStateTool } from './workflows-tools/set-workflow-active-state/set-workflow-active-state-tool';
 import { DeleteWorkflowTool } from './workflows-tools/delete-workflow/delete-workflow-tool';
 
 export const allGraphqlApiTools: BaseMondayApiToolConstructor[] = [
@@ -147,8 +146,7 @@ export const allGraphqlApiTools: BaseMondayApiToolConstructor[] = [
   DeleteAgentTool,
   // Workflows (subgraph still on dev API version)
   ListWorkflowsTool,
-  ActivateWorkflowTool,
-  DeactivateWorkflowTool,
+  SetWorkflowActiveStateTool,
   DeleteWorkflowTool,
 ];
