+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workflows-tools/index.ts
@@ -1,4 +1,3 @@
 export * from './list-workflows/list-workflows-tool';
-export * from './activate-workflow/activate-workflow-tool';
-export * from './deactivate-workflow/deactivate-workflow-tool';
+export * from './set-workflow-active-state/set-workflow-active-state-tool';
 export * from './delete-workflow/delete-workflow-tool';
